refactor
1. updated zap to include caller
2. updated SAM template to include zap logger debug level
3. refactor pkg folder to internal
4. go version updated to 1.20
5 gin version updated to 1.9.1
