+++ .github/workflows/ci-dev.yml
@@ -8,7 +8,7 @@ on:
       - development
     paths:
       - 'lambda/**'
-      - 'pkg/**'
+      - 'internal/**'
       - '.github/workflows/ci-dev.yml'
       - 'tf/development/**'
 
@@ -29,15 +29,15 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v2
       - name: configure aws credentials
-        uses: aws-actions/configure-aws-credentials@v1.6.1
+        uses: aws-actions/configure-aws-credentials@v2.2.0
         with:
           role-to-assume: ${{ secrets.AWS_DEPLOY_ROLE_TO_ASSUME_DEVELOPMENT }}
           role-session-name: githubactionssession
           aws-region: ${{ secrets.AWS_DEFAULT_REGION_DEPLOYMENT_DEVELOPMENT }}
       - name: use golang
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v4
         with:
-          go-version: '^1.17.2'
+          go-version: '^1.20.7'
       - name: Configure git for private modules
         env:
           TOKEN: ${{ secrets.GIT_CLONE_TOKEN }}
